Show installed package versions in settings
Load each project package’s installed version from its generated HISTORY.md and store it on PackageInfo. The HISTORY parser was extracted into a new shared utility (PackageHistoryReader/InstalledPackageReference) so both tooling and workspace code can reuse it. Project Settings now formats and displays the version (e.g. v3) beside the package id when available, with new localized version string support.
