+++ Source/Core/Celbridge.Tools/Helpers/PackageHistoryHelper.cs
@@ -1,13 +1,8 @@
 using System.Globalization;
 using System.Text;
+using Celbridge.Utilities;
 
 namespace Celbridge.Tools;
-
-/// <summary>
-/// The package and version named by the newest HISTORY.md entry, parsed from
-/// its "name@version" heading.
-/// </summary>
-public sealed record InstalledPackageReference(string Name, int Version);
 
 /// <summary>
 /// Formats and parses the generated HISTORY.md changelog written beside a package manifest.
@@ -149,66 +144,21 @@ internal static class PackageHistoryHelper
     }
 
     /// <summary>
-    /// Reads the installed version from a HISTORY.md body. Returns null when the
-    /// file has no parseable version heading (e.g. a hand-authored file).
+    /// Reads the installed version from a HISTORY.md body. Returns null when the file has no parseable
+    /// version heading (e.g. a hand-authored file).
     /// </summary>
     public static int? TryReadInstalledVersion(string historyMarkdown)
     {
-        return TryReadInstalledReference(historyMarkdown)?.Version;
+        return PackageHistoryReader.TryReadInstalledVersion(historyMarkdown);
     }
 
     /// <summary>
-    /// Reads the package and version named by the newest HISTORY.md entry, parsed
-    /// from its "name@version" heading on the first non-empty line. Returns null
-    /// when there is no parseable heading.
+    /// Reads the package and version named by the newest HISTORY.md entry. Returns null when there is no
+    /// parseable heading.
     /// </summary>
     public static InstalledPackageReference? TryReadInstalledReference(string historyMarkdown)
     {
-        if (string.IsNullOrEmpty(historyMarkdown))
-        {
-            return null;
-        }
-
-        using var reader = new StringReader(historyMarkdown);
-        string? line;
-        while ((line = reader.ReadLine()) is not null)
-        {
-            var trimmed = line.Trim();
-            if (trimmed.Length == 0)
-            {
-                continue;
-            }
-
-            // The first non-empty line must be the newest entry's "name@version" heading.
-            var headingText = trimmed.TrimStart('#').Trim();
-
-            // Drop any trailing note after the token.
-            var spaceIndex = headingText.IndexOf(' ');
-            if (spaceIndex > 0)
-            {
-                headingText = headingText.Substring(0, spaceIndex);
-            }
-
-            var atIndex = headingText.LastIndexOf('@');
-            if (atIndex <= 0
-                || atIndex + 1 >= headingText.Length)
-            {
-                return null;
-            }
-
-            var name = headingText.Substring(0, atIndex);
-            var versionText = headingText.Substring(atIndex + 1);
-
-            if (int.TryParse(versionText, NumberStyles.Integer, CultureInfo.InvariantCulture, out var version)
-                && version > 0)
-            {
-                return new InstalledPackageReference(name, version);
-            }
-
-            return null;
-        }
-
-        return null;
+        return PackageHistoryReader.TryReadInstalledReference(historyMarkdown);
     }
 
     /// <summary>
