+++ Source/Core/Celbridge.Tools/Tools/Package/PackageTools.Install.cs
@@ -1,5 +1,6 @@
 using System.Text.Json;
 using Celbridge.Projects;
+using Celbridge.Utilities;
 using ModelContextProtocol.Protocol;
 using ModelContextProtocol.Server;
 using Path = System.IO.Path;
